+++ base/category-source.hpp
@@ -178,7 +178,7 @@ class TypeValue {
   TypeValue() = default;
 
   template<class T>
-  inline BoxedValue Var_self(T* must_be_this) {
+  inline BoxedValue Var_self(T* must_be_this) const {
     if (must_be_this != this) {
       FAIL() << "Var_self called without passing this";
     }